+++ libs/dimcli/cli.cpp
@@ -1,3 +1,6 @@
+// Copyright Glen Knowles 2016 - 2017.
+// Distributed under the Boost Software License, Version 1.0.
+//
 // cli.cpp - dim cli
 //
 // For documentation and examples: