+++ libs/dimcli/cli.h
@@ -160,6 +160,8 @@ public:
     // indirection allows options to be statically registered from multiple
     // source files.
     Cli();
+    Cli(const Cli & from);
+    Cli & operator=(const Cli & from);
 
     //-----------------------------------------------------------------------
     // Configuration