+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-﻿# Change Log
+﻿<!--
+Copyright Glen Knowles 2016 - 2017.
+Distributed under the Boost Software License, Version 1.0.
+-->
+
+# Change Log
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)