+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added - support for clang 3.6
 - Added - support for gcc 6.2
 - Added - show default option values in help text
+- Fixed - cli.toPtrArgv() improperly adding trailing nullptr
 
 ## dimcli 2.0.0 (2017-02-25)
 - Added - optVec.operator[]